% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_build_ml_model.R
\name{nba_build_ml_model}
\alias{nba_build_ml_model}
\title{Build Random Forest Model}
\usage{
nba_build_ml_model(data, target = "points", ntree = 100)
}
\arguments{
\item{data}{A cleaned NBA data frame (output from nba_clean_player_stats)}

\item{target}{A character string specifying the target variable.
Default is "points". Other options include "total_rebounds", "assists", etc.}

\item{ntree}{Number of trees to grow. Default is 100.}
}
\value{
A random forest model object (class "randomForest")
}
\description{
Build a random forest model for predicting NBA player statistics.
This machine learning approach can capture non-linear relationships
and interactions between variables.
}
\examples{
\dontrun{
  data(nba_stats)
  model <- nba_build_ml_model(nba_stats, target = "points")
  predictions <- predict(model, newdata = nba_stats[1:10, ])
}

}
