% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_clean_player_stats.R
\name{nba_clean_player_stats}
\alias{nba_clean_player_stats}
\title{Clean NBA Player Season Statistics}
\usage{
nba_clean_player_stats(data)
}
\arguments{
\item{data}{A data frame containing raw NBA player season statistics.}
}
\value{
A cleaned data frame.
}
\description{
This function cleans NBA player season statistics by removing
unnecessary columns and renaming important variables.
}
