% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_build_ml_model.R
\name{nba_compare_models}
\alias{nba_compare_models}
\title{Compare Statistical and Machine Learning Models}
\usage{
nba_compare_models(stat_model, ml_model, test_data)
}
\arguments{
\item{stat_model}{A linear regression model (output from nba_build_regression_model)}

\item{ml_model}{A random forest model (output from nba_build_ml_model)}

\item{test_data}{Test data for evaluation}
}
\value{
A list containing comparison metrics:
\item{stat_rmse}{RMSE of statistical model}
\item{ml_rmse}{RMSE of machine learning model}
\item{stat_mae}{MAE of statistical model}
\item{ml_mae}{MAE of machine learning model}
\item{stat_r2}{R-squared of statistical model}
\item{ml_r2}{R-squared of machine learning model}
}
\description{
Compare the performance of linear regression and random forest models
using various evaluation metrics (RMSE, MAE, R-squared).
}
\details{
RMSE (Root Mean Squared Error): Lower is better
MAE (Mean Absolute Error): Lower is better
R-squared: Higher is better (0-1 scale)
}
\examples{
\dontrun{
  data(nba_stats)
  stat_model <- nba_build_regression_model(nba_stats)
  ml_model <- nba_build_ml_model(nba_stats)
  test_data <- nba_stats \%>\% dplyr::slice(1:500)
  comparison <- nba_compare_models(stat_model, ml_model, test_data)
  print(comparison)
}

}
