% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_load_player_stats.R
\name{nba_load_player_stats}
\alias{nba_load_player_stats}
\title{Load NBA Player Season Statistics}
\usage{
nba_load_player_stats(file_path = "data-raw/NBA Player Stats(1950 - 2022).csv")
}
\arguments{
\item{file_path}{A character string giving the path to the CSV file.}
}
\value{
A data frame containing NBA player season statistics.
}
\description{
This function reads NBA player season statistics from a CSV file.
}
